+++ libraries/DAC8571/DAC8571.cpp
@@ -2,7 +2,7 @@
 //    FILE: DAC8571.cpp
 //  AUTHOR: Rob Tillaart
 //    DATE: 2024-04-16
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 // PURPOSE: Arduino library for DAC8571 I2C.
 //     URL: https://github.com/RobTillaart/DAC8571
 
@@ -25,7 +25,7 @@ DAC8571::DAC8571(uint8_t address, TwoWire *wire)
 }
 
 
-bool DAC8571::begin(uint16_t val)
+bool DAC8571::begin(uint16_t value)
 {
   if ((_address != 0x4C) && (_address != 0x4E))
   {
@@ -34,7 +34,7 @@ bool DAC8571::begin(uint16_t val)
   }
   if (!isConnected()) return false;
 
-  return write(val);
+  return write(value);
 }
 
 
@@ -42,7 +42,12 @@ bool DAC8571::isConnected()
 {
   _wire->beginTransmission(_address);
   _error = _wire->endTransmission();  //  default == 0 ==> DAC8571_OK
-  return (_error == DAC8571_OK);
+  if (_error != 0)
+  {
+    _error = DAC8571_I2C_ERROR;
+    return false;
+  }
+  return true;
 }
 
 
@@ -86,23 +91,15 @@ uint16_t DAC8571::lastWrite()
 
 uint16_t DAC8571::read()
 {
-  uint8_t highByte = 0;
-  uint8_t lowByte  = 0;
-  //  uint8_t control  = 0;  //  not used.
-
   uint8_t n = _wire->requestFrom(_address, uint8_t(3));
-  if (n == 3)
-  {
-    highByte = _wire->read();
-    lowByte  = _wire->read();
-    _wire->read();    //  control  = _wire->read();  //  not used.
-    _error = DAC8571_OK;
-  }
-  else
+  if (n != 3)
   {
     _error = DAC8571_I2C_ERROR;
     return 0;
   }
+  uint8_t highByte = _wire->read();
+  uint8_t lowByte  = _wire->read();
+  _wire->read();    //  uint8_t control  = _wire->read();  //  not used.
   _error = DAC8571_OK;
   uint16_t value = highByte * 256 + lowByte;
   return value;
@@ -116,6 +113,7 @@ bool DAC8571::write(uint16_t * arr, uint8_t length)
     _error = DAC8571_BUFFER_ERROR;
     return false;
   }
+
   _wire->beginTransmission(_address);
   _wire->write(_control);
   for (int i = 0; i < length; i++)
@@ -126,12 +124,12 @@ bool DAC8571::write(uint16_t * arr, uint8_t length)
     _wire->write(lowByte);
   }
   _error = _wire->endTransmission();
-
   if (_error != 0)
   {
     _error = DAC8571_I2C_ERROR;
     return false;
   }
+
   _error = DAC8571_OK;
   _dac = arr[length - 1];
   return true;
@@ -142,17 +140,19 @@ bool DAC8571::write(uint16_t * arr, uint8_t length)
 //
 //  PERCENTAGE WRAPPER
 //
-void DAC8571::setPercentage(float percentage)
+bool DAC8571::setPercentage(float percentage)
 {
   if (percentage < 0) percentage = 0;
   else if (percentage > 100) percentage = 100;
-  write(percentage * 655.35);
+  return write(percentage * 655.35);
 }
 
 
 float DAC8571::getPercentage()
 {
-  return read() * 0.0015259022;  //  === / 655.35;
+  //  return read() * 0.0015259022;  //  === / 655.35;
+  //  faster from cache
+  return _dac * 0.0015259022;  //  === / 655.35;
 }
 
 
@@ -162,8 +162,7 @@ float DAC8571::getPercentage()
 //
 void DAC8571::setWriteMode(uint8_t mode)
 {
-  //  broadcast modi (3,4,5) not supported.
-  if (mode > DAC8571_MODE_WRITE_CACHE)
+  if (mode > DAC8571_MODE_BRCAST_0)
   {
     mode = DAC8571_MODE_NORMAL;
   }
@@ -182,7 +181,7 @@ uint8_t DAC8571::getWriteMode()
 //
 //  POWER DOWN PART
 //
-void DAC8571::powerDown(uint8_t pdMode)
+bool DAC8571::powerDown(uint8_t pdMode)
 {
   uint16_t pdMask = 0x0000;
   //  table 6, page 22.
@@ -205,16 +204,16 @@ void DAC8571::powerDown(uint8_t pdMode)
       pdMask  = 0xC000;
       break;
   }
-  //  specific power down code.
-  _control = 0x11;
-  write(pdMask);
+  //  DAC8571_MODE_NORMAL + power down bit 7
+  _control = (DAC8571_MODE_NORMAL << 4) + 0x01;
+  return write(pdMask);
 }
 
 
-void DAC8571::wakeUp(uint16_t value)
+bool DAC8571::wakeUp(uint16_t value)
 {
   setWriteMode(DAC8571_MODE_NORMAL);
-  write(value);
+  return write(value);
 }