+++ libraries/DAC8571/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2024-11-14
+- fix #4, sync with DAC8574
+- change return type **bool powerDown()**
+- change return type **bool wakeUp()**
+- change return type **bool setPercentage()**
+- change **getPercentage()** to return cached value. faster than read().
+- enable **DAC8571_MODE_BRCAST_0** (experimental)
+- update keywords.txt
+- minor edits.
+
 ## [0.1.1] - 2024-04-18
 - implement more powerDown modi
 - update examples